+++ packages/components/eslint.config.mjs
@@ -11,7 +11,7 @@ import storybookPlugin from "eslint-plugin-storybook";
 
 export default tseslint.config(
   {
-    ignores: ["**/dist", "**/docs", "**/hydrate", "**/*.d.ts"],
+    ignores: ["**/dist", "**/docs", "**/*.d.ts"],
   },
 
   {
@@ -23,7 +23,11 @@ export default tseslint.config(
     },
 
     languageOptions: {
+      ecmaVersion: 2021,
+      sourceType: "module",
+      parser: tseslint.parser,
       parserOptions: {
+        jsxPragma: "h",
         tsconfigRootDir: import.meta.dirname,
         project: ["tsconfig.eslint.json"],
       },
@@ -44,7 +48,7 @@ export default tseslint.config(
             {
               group: ["tests/commonTests/browser/*"],
               message:
-                "Import named functions from commonTests/browser for browser mode (experimental) tests instead of direct module imports, e.g., import { cancelable } from 'tests/commonTests/browser'",
+                "Import named functions from commonTests/browser for browser mode tests instead of direct module imports, e.g., import { cancelable } from 'tests/commonTests/browser'",
             },
             {
               group: ["lit-html", "lit-html/*"],