+++ packages/components/src/components/card/card.tsx
@@ -10,8 +10,9 @@ import {
   state,
   ToEvents,
 } from "@arcgis/lumina";
+import { getIconScale } from "../../utils/component";
 import { slotChangeHasAssignedElement } from "../../utils/dom";
-import { LogicalFlowPosition, SelectionMode } from "../interfaces";
+import { LogicalFlowPosition, Scale, SelectionMode } from "../interfaces";
 import { isActivationKey } from "../../utils/key";
 import { IconName } from "../icon/interfaces";
 import { useT9n } from "../../controllers/useT9n";
@@ -89,6 +90,11 @@ export class Card extends LitElement {
 
   /** Overrides individual strings used by the component. */
   @property() messageOverrides?: typeof this.messages._overrides;
+
+  /**
+   * Specifies the size of the component. When contained in a parent `calcite-card-group`, inherits the parent's `scale` value, but can be overridden if needed.
+   */
+  @property({ reflect: true }) scale: Scale = "m";
 
   /**
    * When `true`, the component is selectable.
@@ -246,7 +252,7 @@ export class Card extends LitElement {
 
     return (
       <div class={CSS.checkboxWrapper} onPointerDown={this.cardSelectClick} tabIndex={-1}>
-        <calcite-icon icon={icon} scale="s" />
+        <calcite-icon icon={icon} scale={getIconScale(this.scale)} />
       </div>
     );
   }