+++ packages/components/src/components/input-date-picker/input-date-picker.browser.e2e.tsx
@@ -1,9 +1,8 @@
 import { h } from "@arcgis/lumina";
-import { describe } from "vitest";
+import { describe, it, expect } from "vitest";
 import { mount } from "@arcgis/lumina-compiler/testing";
 import {
   defaults,
-  disabled,
   focusable,
   hidden,
   internalLabel,
@@ -16,6 +15,7 @@ import {
 } from "../../tests/commonTests/browser";
 import { mockConsole } from "../../tests/utils/logging";
 import { defaultValidity } from "../../tests/commonTests/browser/defaults";
+import { InputDatePicker } from "./input-date-picker";
 
 describe("defaults", () => {
   defaults(
@@ -93,9 +93,7 @@ describe("translation support", () => {
   t9n(() => mount("calcite-input-date-picker"));
 });
 
-describe.skip("disabled", () => {
-  disabled(() => mount("calcite-input-date-picker"));
-});
+describe.todo("disabled");
 
 describe("is form-associated", () => {
   mockConsole();
@@ -118,5 +116,25 @@ describe("is form-associated", () => {
         inputType: "date",
       },
     );
+  });
+});
+
+describe("minAsDate and maxAsDate properties", () => {
+  it("honors minAsDate and maxAsDate properties by updating out-of-range value to the closest valid value", async () => {
+    const { el, component } = await mount<InputDatePicker>(
+      <calcite-input-date-picker value="2022-11-27" />,
+    );
+
+    const offsetTime = `T09:00:00.000Z`;
+    el.minAsDate = new Date(`2020-01-01${offsetTime}`);
+    el.maxAsDate = new Date(`2020-12-31${offsetTime}`);
+    await component.updateComplete;
+
+    expect(el.value).toBe("2020-12-31");
+
+    const input = el.shadowRoot
+      .querySelector<HTMLElement>("calcite-input-text")
+      ?.shadowRoot.querySelector<HTMLInputElement>("input");
+    expect(input.value).toBe("12/31/2020");
   });
 });