+++ packages/components/src/components/radio-button/radio-button.browser.e2e.tsx
@@ -1,7 +1,9 @@
 import { describe } from "vitest";
+import { h } from "@arcgis/lumina";
 import { mount } from "@arcgis/lumina-compiler/testing";
 import {
   focusable,
+  formAssociated,
   internalLabel,
   renders,
   disabled,
@@ -9,9 +11,58 @@ import {
   reflects,
   hidden,
 } from "../../tests/commonTests/browser";
+import { defaultValidity } from "../../tests/commonTests/browser/defaults";
 
 describe("defaults", () => {
-  defaults(() => mount("calcite-radio-button"), [{ propertyName: "scale", defaultValue: "m" }]);
+  defaults(
+    () => mount("calcite-radio-button"),
+    [
+      { propertyName: "scale", defaultValue: "m" },
+      {
+        propertyName: "validity",
+        defaultValue: defaultValidity,
+      },
+    ],
+  );
+});
+
+describe("is form-associated", () => {
+  describe("no group", () => {
+    formAssociated(() => mount("calcite-radio-button"), {
+      testValue: true,
+      inputType: "radio",
+    });
+  });
+
+  // skipped until the util supports a parent component wrapping the form associated element(s)
+  // https://github.com/Esri/calcite-design-system/issues/9221
+  describe.skip("group", () => {
+    formAssociated(
+      () =>
+        mount(
+          <calcite-radio-button-group name="using" required>
+            <calcite-label layout="inline">
+              Yes
+              <calcite-radio-button required value="yes" />
+            </calcite-label>
+            <calcite-label layout="inline">
+              No
+              <calcite-radio-button required value="no" />
+            </calcite-label>
+            <calcite-label layout="inline">
+              Maybe
+              <calcite-radio-button required value="maybe" />
+            </calcite-label>
+          </calcite-radio-button-group>,
+        ),
+      {
+        testValue: true,
+        inputType: "radio",
+        validation: true,
+        changeValueKeys: ["{Space}"],
+      },
+    );
+  });
 });
 
 describe("reflects", () => {