+++ packages/eslint-plugin-components/tests/lib/rules/require-deprecation-and-removal-versions/require-deprecation-and-removal-version.spec.ts
@@ -1,4 +1,3 @@
-import { describe } from "vitest";
 import rule from "../../../../src/rules/require-deprecation-details";
 import good from "./require-deprecation-and-removal-version.good.tsx?raw";
 import wrongMissingDeprecation from "./require-deprecation-version.wrong.tsx?raw";